Serialize bounds2

+++ core/src/main/java/org/graphast/query/route/osr/OSRSearch.java
@@ -208,7 +208,7 @@ public class OSRSearch {
 		if(pos < categories.size()){
 			int distance;
 			for(int i = pos; i < categories.size(); i++){
-				distance = bounds.getBound(id, categories.get(i)).getCost();
+				distance = bounds.getBound(id, (int)categories.get(i)).getCost();
 				if(distance > max)	max = distance;
 			}
 		}
